requirements.txt
pyautogen==0.2.18
arxiv==2.1.0
PyPDF2==3.0.1
requests==2.31.0
openai==1.3.8
pillow==10.2.0
fastapi==0.110.0
uvicorn==0.27.0.post1
fastapi-healthchecks==1.1.0

# psycopg2==2.9.9